backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
openai>=1.6.1,<2.0.0
aiohttp==3.9.1
httpx==0.25.2
python-multipart==0.0.6

